+++ packages/rainbowkit/src/components/RainbowKitProvider/provideRainbowKitChains.ts
@@ -28,10 +28,12 @@ type ChainName =
   | 'rinkeby'
   | 'ropsten'
   | 'sepolia'
-  | 'zora'
-  | 'zoraTestnet'
+  | 'xdc'
+  | 'xdcTestnet'
   | 'zkSync'
-  | 'zkSyncTestnet';
+  | 'zkSyncTestnet'
+  | 'zora'
+  | 'zoraTestnet';
 
 type IconMetadata = {
   iconUrl: () => Promise<string>;
@@ -88,14 +90,19 @@ const polygonIcon: IconMetadata = {
   iconUrl: async () => (await import('./chainIcons/polygon.svg')).default,
 };
 
-const zoraIcon: IconMetadata = {
-  iconBackground: '#000000',
-  iconUrl: async () => (await import('./chainIcons/zora.svg')).default,
+const xdcIcon: IconMetadata = {
+  iconBackground: '#f9f7ec',
+  iconUrl: async () => (await import('./chainIcons/xdc.svg')).default,
 };
 
 const zkSyncIcon: IconMetadata = {
   iconBackground: '#f9f7ec',
   iconUrl: async () => (await import('./chainIcons/zkSync.svg')).default,
+};
+
+const zoraIcon: IconMetadata = {
+  iconBackground: '#000000',
+  iconUrl: async () => (await import('./chainIcons/zora.svg')).default,
 };
 
 const chainMetadataByName: Record<ChainName, ChainMetadata | null> = {
@@ -123,10 +130,12 @@ const chainMetadataByName: Record<ChainName, ChainMetadata | null> = {
   rinkeby: { chainId: 4, ...ethereumIcon },
   ropsten: { chainId: 3, ...ethereumIcon },
   sepolia: { chainId: 11_155_111, ...ethereumIcon },
-  zora: { chainId: 7777777, name: 'Zora', ...zoraIcon },
-  zoraTestnet: { chainId: 999, ...zoraIcon },
+  xdc: { chainId: 50, name: 'XinFin', ...xdcIcon },
+  xdcTestnet: { chainId: 51, ...xdcIcon },
   zkSync: { chainId: 324, name: 'zkSync', ...zkSyncIcon },
   zkSyncTestnet: { chainId: 280, ...zkSyncIcon },
+  zora: { chainId: 7777777, name: 'Zora', ...zoraIcon },
+  zoraTestnet: { chainId: 999, ...zoraIcon },
 };
 
 const chainMetadataById = Object.fromEntries(