+++ examples/with-create-react-app/CHANGELOG.md
@@ -1,5 +1,14 @@
 # with-create-react-app
 
+## 0.1.71
+
+### Patch Changes
+
+- Updated dependencies [74ead9df]
+- Updated dependencies [94dce820]
+- Updated dependencies [39d81e93]
+  - @rainbow-me/rainbowkit@1.2.1
+
 ## 0.1.70
 
 ### Patch Changes