+++ examples/with-vite/CHANGELOG.md
@@ -1,5 +1,14 @@
 # with-vite
 
+## 0.0.59
+
+### Patch Changes
+
+- Updated dependencies [74ead9df]
+- Updated dependencies [94dce820]
+- Updated dependencies [39d81e93]
+  - @rainbow-me/rainbowkit@1.2.1
+
 ## 0.0.58
 
 ### Patch Changes