+++ packages/example/CHANGELOG.md
@@ -1,5 +1,15 @@
 # example
 
+## 0.0.81
+
+### Patch Changes
+
+- Updated dependencies [74ead9df]
+- Updated dependencies [94dce820]
+- Updated dependencies [39d81e93]
+  - @rainbow-me/rainbowkit@1.2.1
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.3.2
+
 ## 0.0.80
 
 ### Patch Changes